Rearrange files to better match the Getting Started Guide
Add source header template and apply standard template to files

Update marketing and project verions

Set deployment target to iOS 15
