+++ swift-spaces-real-time-app/MuxSpacesSamples.xcodeproj/xcshareddata/IDETemplateMacros.plist
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>FILEHEADER</key>
+	<string>
+//  Created for ___PACKAGENAME___.
+//  
+//  ___NSHUMANREADABLECOPYRIGHTPLIST___
+//  Licensed under the MIT License.
+//</string>
+	<key>NSHUMANREADABLECOPYRIGHTPLIST</key>
+	<string>Copyright © ___YEAR___ Mux, Inc.</string>
+</dict>
+</plist>
