+++ swift-spaces-real-time-app/MuxSpacesUIKit/AppDelegate.swift
@@ -1,5 +1,8 @@
 //
-//  AppDelegate.swift
+//  Created for MuxSpacesUIKit.
+//
+//  Copyright © 2022 Mux, Inc.
+//  Licensed under the MIT License.
 //
 
 import UIKit
