+++ swift-spaces-real-time-app/MuxSpacesUIKit/ExampleUsingNavigation/JoinSpaceViewController.swift
@@ -1,6 +1,9 @@
 //
-//  JoinSpaceViewController.swift
-//  
+//  Created for MuxSpacesUIKit.
+//
+//  Copyright © 2022 Mux, Inc.
+//  Licensed under the MIT License.
+//
 
 import Combine
 import UIKit
