--- swift-spaces-real-time-app/MuxSpacesUIKit/Model/ParticipantsSnapshot+Make.swift
@@ -1,15 +1,0 @@
-//
-//  ParticipantsSnapshot+Make.swift
-//
-
-import UIKit
-
-extension ParticipantsSnapshot {
-    static func make() -> Self {
-        var snapshot = ParticipantsSnapshot()
-        snapshot.appendSections(
-            [.participants]
-        )
-        return snapshot
-    }
-}
