+++ swift-spaces-real-time-app/MuxSpacesUIKit/Shared/ParticipantLayout.swift
@@ -1,5 +1,8 @@
 //
-//  ParticipantLayout.swift
+//  Created for MuxSpacesUIKit.
+//
+//  Copyright © 2022 Mux, Inc.
+//  Licensed under the MIT License.
 //
 
 import AVFoundation
