+++ swift-spaces-real-time-app/MuxSpacesUIKit/Shared/TargetAction.swift
@@ -1,5 +1,8 @@
 //
-//  TargetAction.swift
+//  Created for MuxSpacesUIKit.
+//
+//  Copyright © 2022 Mux, Inc.
+//  Licensed under the MIT License.
 //
 
 import Foundation
