+++ swift-spaces-real-time-app/MuxSpacesUIKit/Shared/UserDefaultsKeys.swift
@@ -1,11 +1,14 @@
 //
-//  UserDefaultsKeys.swift
-//  
+//  Created for MuxSpacesUIKit.
+//
+//  Copyright © 2022 Mux, Inc.
+//  Licensed under the MIT License.
+//
 
 import Foundation
 
 enum UserDefaultsKeys {
 
-    static let customParticipantID = "com.mux.spaces.test.customParticipantID"
+    static let customParticipantID = "com.mux.spaces.demonstration.customParticipantID"
 
 }
