+++ swift-spaces-real-time-app/MuxSpacesUIKit/SpaceSetup/ProcessInfo+EnvironmentVariables.swift
@@ -1,5 +1,8 @@
 //
-//  ProcessInfo+EnvironmentVariables.swift
+//  Created for MuxSpacesUIKit.
+//
+//  Copyright © 2022 Mux, Inc.
+//  Licensed under the MIT License.
 //
 
 import Foundation
