+++ swift-spaces-real-time-app/MuxSpacesUIKitUITests/MuxSpacesUIKitUITests.swift
@@ -1,5 +1,8 @@
 //
-//  MuxSpacesUIKitUITests.swift
+//  Created for MuxSpacesUIKitUITests.
+//
+//  Copyright © 2022 Mux, Inc.
+//  Licensed under the MIT License.
 //
 
 import XCTest
